requirements.txt
PyAudio
websocket_client